refactor: 인수테스트 분리 - product
