+++ docs/kitchenpose.md
@@ -112,6 +112,10 @@
 ### 4단계 
 - [ ] 인수테스트 분리
   - [X] menu
+  - [X] product
+  - [ ] table
+  - [ ] order
+
 - [ ] 모듈 생성
   - module-core
   - module-product
