+++ src/test/java/kitchenpos/product/acceptance/ProductAcceptanceTest.java
@@ -7,6 +7,7 @@ import io.restassured.response.Response;
 import java.util.List;
 import kitchenpos.AcceptanceTest;
 import kitchenpos.acceptance.util.KitchenPosBehaviors;
+import kitchenpos.product.acceptance.behavior.ProductContextBehavior;
 import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,10 +17,10 @@ class ProductAcceptanceTest extends AcceptanceTest {
     @Test
     @DisplayName("상품 생성 및 조회 기능 인수테스트")
     void productAcceptanceTest() {
-        ExtractableResponse<Response> createResponse = KitchenPosBehaviors.상품_생성_요청("강정치킨", 17000);
+        ExtractableResponse<Response> createResponse = ProductContextBehavior.상품_생성_요청("강정치킨", 17000);
         assertThat(createResponse.statusCode()).isEqualTo(HttpStatus.CREATED.value());
 
-        List<Product> products = KitchenPosBehaviors.상품목록_조회();
+        List<Product> products = ProductContextBehavior.상품목록_조회();
         assertThat(products).hasSize(1);
     }
 }
